Finish collision
- finish collision test between shape !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
- change some code duplication by reusable functions (collisionUtils.h)
- maybe change the anonymous namespace by something else ...

+++ Sources/Events/EventHandlerString.cpp
@@ -319,7 +319,7 @@ unsigned int EventHandlerString::getNumberOfEvent()
 {
 	unsigned int nb;
 	mutex.lock();
-	nb = keyboardListeners.size();
+	nb = (unsigned int) keyboardListeners.size();
 	mutex.unlock();
 	return nb;
 }
